# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/{hotelId}/rooms": {
      "get": {
        "tags": [
          "Hotels"
        ],
        "description": "Retrieve all rooms available for a specific hotel.",
        "produces": [
          "text/html"
        ],
        "parameters": [
          {
            "name": "hotelId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the hotel"
          }
        ],
        "responses": {
          "200": {
            "description": "Rooms retrieved successfully.",
            "content": {
              "text/html": {}
            }
          },
          "404": {
            "description": "Hotel not found or has no rooms.",
            "content": {
              "text/html": {
                "schema": {
                  "title": "Error",
                  "message": "Hotel Not Found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "text/html": {
                "schema": {
                  "title": "Error",
                  "message": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/{hotelId}/reservations": {
      "get": {
        "tags": [
          "Hotels"
        ],
        "description": "Retrieve all reservations for a specific hotel.",
        "produces": [
          "text/html"
        ],
        "parameters": [
          {
            "name": "hotelId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Hotel ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Reservations retrieved successfully.",
            "content": {
              "text/html": {}
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/rooms/add": {
      "post": {
        "tags": [
          "Rooms"
        ],
        "description": "Create a new room (Admin Only).",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Room details to create.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Room"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Room added successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "message": "✅ Room added successfully!"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Missing required fields.",
            "content": {
              "application/json": {
                "schema": {
                  "message": "❌ Capacity, price, and hotelId are required!"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error - Failed to add room.",
            "content": {
              "application/json": {
                "schema": {
                  "message": "❌ Internal server error."
                }
              }
            }
          }
        }
      }
    },
...
    "/reservations/{reservationId}": {
      "delete": {
        "tags": [
          "Reservations"
        ],
        "description": "Cancel an existing reservation by its ID.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "reservationId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the reservation to be deleted."
          }
        ],
        "responses": {
          "200": {
            "description": "Reservation canceled successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "message": "✅ Reservation canceled successfully!"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Reservation ID is required.",
            "content": {
              "application/json": {
                "schema": {
                  "message": "Reservation ID is required."
                }
              }
            }
          },
          "404": {
            "description": "Reservation not found.",
            "content": {
              "application/json": {
                "schema": {
                  "message": "Reservation not found."
                }
              }
            }
          },
          "500": {
            "description": "Internal server error - Failed to cancel reservation.",
            "content": {
              "application/json": {
                "schema": {
                  "message": "Failed to cancel reservation."
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Reservations"
        ],
        "description": "Retrieve details of a specific reservation.",
        "produces": [
          "text/html"
        ],
        "parameters": [
          {
            "name": "reservationId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the reservation to retrieve."
          }
        ],
        "responses": {
          "200": {
            "description": "Reservation details retrieved successfully.",
            "content": {
              "text/html": {
                "schema": {
                  "title": "Reservation Details",
                  "reservation": {
                    "id": 1,
                    "user": {
                      "id": 3,
                      "name": "John Doe",
                      "email": "john@example.com"
                    },
                    "hotel": {
                      "id": 2,
                      "name": "Grand Hotel",
                      "location": "New York"
                    },
                    "room": {
                      "id": 5,
                      "type": "Suite",
                      "price": 200
                    },
                    "startDate": "2025-03-20",
                    "endDate": "2025-03-25"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Reservation not found.",
            "content": {
              "text/html": {
                "schema": {
                  "title": "Error",
                  "message": "Reservation Not Found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error - Failed to fetch reservation details.",
            "content": {
              "text/html": {
                "schema": {
                  "title": "Error",
                  "message": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    },
    "/reservations": {
      "post": {
        "tags": [
          "Reservations"
        ],
        "description": "Create a new reservation (Only authorized users).",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Room reservation details.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "number",
                  "example": 1
                },
                "roomId": {
                  "type": "number",
                  "example": 5
                },
                "startDate": {
                  "type": "string",
                  "example": "2025-03-20"
                },
                "endDate": {
                  "type": "string",
                  "example": "2025-03-25"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Room reserved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "success": true,
                  "message": "✅ Room reserved successfully!"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid reservation data.",
            "content": {
              "application/json": {
                "schema": {
                  "message": "❌ Invalid request data."
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User must be logged in."
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error - Failed to reserve room.",
            "content": {
              "application/json": {
                "schema": {
                  "message": "❌ Failed to reserve room."
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Room": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "capacity": {
          "type": "number",
          "example": 2
        },
        "price": {
          "type": "number",
          "example": 100
        },
        "hotelId": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "id",
        "capacity",
        "price",
        "hotelId"
      ]
    },
...
  }
}